VM extension refactor
